Update motif columns by data values

**NOTE** ignores NA values

mutate_motif(.data, ..., .motif = "motif", .override = NULL)

Arguments

.data
data.frame of results

...
quoted name-value pairs of columns to edit. format: "motif_column_name" = "data_column_name"

.motif
column of universalmotif objects to edit (default: "motif")

.override
named list or character vector where names are motif_column_name, values are data_column_name, used instead of ... (useful for programatically passing values)

Value

`.data` where `.motif` entries have been updated to values passed to `...` or `.override`

Details

**NOTE** This feature is experimental and subject to change based on user-feedback. Please provide feedback at https://github.com/snystrom/dremeR/issues/31

Examples

motif <- universalmotif::create_motif()

#> Registered S3 method overwritten by 'treeio':
#>   method     from
#>   root.phylo ape 

df <- universalmotif_to_meme_df(motif)

#> Error in universalmotif_to_meme_df(motif): could not find function "universalmotif_to_meme_df"

df <- dplyr::mutate(df, id = "newName")

#> Error in UseMethod("mutate_"): no applicable method for 'mutate_' applied to an object of class "function"

df <- dplyr::mutate(df, name = "newName")

#> Error in UseMethod("mutate_"): no applicable method for 'mutate_' applied to an object of class "function"

df <- mutate_motif(df, "name" = "name")

#> Error in mutate_motif(df, name = "name"): .motif %in% names(.data) is not TRUE

# renamed motif
df$motif

#> Error in df$motif: object of type 'closure' is not subsettable