Motif enrichment using AME

Motif enrichment using AME

runAme(
  input,
  control = "shuffle",
  outdir = "auto",
  method = "fisher",
  database = NULL,
  meme_path = NULL,
  sequences = FALSE,
  silent = TRUE,
  ...
)

Arguments

input
path to fasta, or DNAstringset (optional: DNAStringSet object names contain fasta score, required for partitioning mode)

control
default: "shuffle", or set to DNAstringset or path to fasta file to use those sequences in discriminative mode. If `input` is a list of DNAStringSet objects, and `control` is set to a character vector of names in `input`, those regions will be used as background in discriminitive mode and AME will skip running on any `control` entries (NOTE: if `input` contains an entry named "shuffle" and control is set to "shuffle", it will use the `input` entry, not the AME shuffle algorithm). If `control` is a Biostrings::BStringSetList (generated using `get_sequence`), all sequences in the list will be combined as the control set. Set to `NA` for partitioning based on input fasta score (see `get_sequence()` for assigning fasta score).

outdir
default: auto

method
default: fisher

database
path to .meme format file, universalmotif list object, dreme results data.frame, or list() of multiple of these. If objects are assigned names in the list, that name will be used as the database id. It is highly recommended you set a name if not using a file path so the database name will be informative, otherwise the position in the list will be used as the database id. For example, if the input is: list("motifs.meme", list_of_motifs), the database id's will be: "motifs.meme" and "2". If the input is list("motifs.meme", "customMotifs" = list_of_motifs), the database id's will be "motifs.meme" and "customMotifs".

meme_path
path to "meme/bin/" (default: `NULL`). Will use default search behavior as described in `check_meme_install()` if unset.

sequences
`logical(1)` add results from `sequences.tsv` to `sequences` list column to returned data.frame. Valid only if method = "fisher". See [AME outputs](http://alternate.meme-suite.org/doc/ame-output-format.html) webpage for more information.

silent
whether to suppress stdout (default: TRUE), useful for debugging.

...

Value

Examples

if (FALSE) {
dreme_out <- runDreme("input.fa")
runAme("input.fa", database = "jasparmotifs.meme")
runAme("input.fa", database = list("jasparmotifs.meme", "my_dreme_motifs" = dreme_out))
}