Denovo motif discovery of target regions

Denovo motif discovery of target regions

runDreme(input, control, outdir = "auto", meme_path = NULL, silent = TRUE, ...)

Arguments

input
regions to scan for motifs. Can be any of:

path to fasta file

DNAStringSet object (can be generated from GRanges using get_sequence)

NOTE: If you want to retain the raw dreme output files, you must use a path to fasta file as input, or specify an "outdir"

control
regions to use as background for motif search. Can be any of:

path to fasta file

DNAStringSet object (can be generated from GRanges using get_sequence)

"shuffle" to use dreme's built-in dinucleotide shuffle feature. Optionally can also pass s = <any number> to ... to use as the random seed during shuffling. If no seed is passed, dreme will use 1 as the random seed, so results will be reproducible if rerunning. NOTE: beware system-specific differences. As of v5, dreme will compile using the default python installation on a system (either python2.7 or python3). The random number generator changed between python2.7 and python3, so results will not be reproducible between systems using python2.7 vs 3.

outdir
path to output directory of dreme files, or "auto" to autogenerate path. Default: location of input fasta in dir named "\<input\>vs\<control\>". If input is DNAstringset, will be temporary path. This means that if you want to save the raw output files, you must use fasta files as input or use an informative (and unique) outdir name. dremeR will not check if it overwrites files in a directory. Directories will be recursively created if needed.

meme_path
optional, path to "meme/bin/" on your local machine. runDreme will search 3 places in order for meme if this flag is unset:

the environment variable "MEME_PATH" (set in .Renviron)

the option "meme_bin" (set with options(meme_bin = "path/to/meme/bin"))

"~/meme/bin/" as the default location

If the user sets meme_path in the function call, this value will always be used

silent
whether to suppress printing dreme stdout as a message when finishing with no errors. Can be useful for troubleshooting in situations where no motifs are discovered, but command completes successfully. (default: TRUE)

...
dreme flags can be passed as R function arguments to use non-default behavior. For a full list of valid arguments, run your local install of dreme -h, or visit the dreme documentation website.

Value

data.frame with statistics for each discovered motif. The motif column contains a universalmotif object representation in PCM format of each DREME motif. If no motifs are discovered, returns NULL.

Details

In addition to allowing any valid flag of dreme to be passed to ..., we provide a few user-friendly aliases for common flags which will hopefully make it things easier to remember. For example, m = 2 will search for 2 motifs. This is equivalent to setting nmotifs = 2.

List of aliased values which can be passed to ...

nmotifs = max number of motifs to search for

sec = max runtime in seconds

evalue = max evalue cutoff

seed = random seed if using "shuffle" as control

ngen = number of REs to generalize

Examples

if (FALSE) {
# Runs dreme with default settings, shuffles input as background
runDreme("input.fa", "shuffle")

# Runs searching for max 2 motifs, e-value cutoff = 0.1, explicitly using the DNA alphabet
runDreme("input.fa", "shuffle", m = 2, e = 0.1, dna = T)
}