Update motif columns by data values

**NOTE** ignores NA values

mutate_motif(.data, ..., .motif = "motif", .override = NULL)

Arguments

.data
data.frame of results

...
quoted name-value pairs of columns to edit. format: "motif_column_name" = "data_column_name"

.motif
column of universalmotif objects to edit (default: "motif")

.override
named list or character vector where names are motif_column_name, values are data_column_name, used instead of ... (useful for programatically passing values)

Value

`.data` where `.motif` entries have been updated to values passed to `...` or `.override`

Details

**NOTE** This feature is experimental and subject to change based on user-feedback. Please provide feedback at https://github.com/snystrom/dremeR/issues/31

Examples

motif <- universalmotif::create_motif()
df <- as_universalmotif_dataframe(motif)
df <- dplyr::mutate(df, id = "newName")
df <- mutate_motif(df, "name" = "id")
# renamed motif
df$motif

#> [[1]]
#> 
#>        Motif name:   newName
#>          Alphabet:   DNA
#>              Type:   PPM
#>           Strands:   +-
#>          Total IC:   8.53
#>         Consensus:   RRTNGGARTN
#> 
#>      R   R    T    N    G    G    A    R    T    N
#> A 0.52 0.3 0.03 0.37 0.00 0.02 0.77 0.41 0.14 0.22
#> C 0.00 0.0 0.11 0.29 0.02 0.13 0.00 0.05 0.00 0.23
#> G 0.48 0.7 0.00 0.23 0.85 0.62 0.06 0.51 0.00 0.25
#> T 0.00 0.0 0.86 0.12 0.13 0.23 0.17 0.02 0.86 0.30
#>